+++ src-ui/app/services/brightness-control/simple-brightness-control.service.ts
@@ -2,22 +2,23 @@ import { Injectable } from '@angular/core';
 import { BehaviorSubject, distinctUntilChanged, filter, map, Observable, skip, tap } from 'rxjs';
 import { info } from 'tauri-plugin-log-api';
 import { CancellableTask } from '../../utils/cancellable-task';
-import { createBrightnessTransitionTask } from './brightness-transition';
+import { BrightnessTransitionTask } from './brightness-transition';
 import { AutomationConfigService } from '../automation-config.service';
 import { DisplayBrightnessControlService } from './display-brightness-control.service';
 import { ImageBrightnessControlService } from './image-brightness-control.service';
 import { lerp } from '../../utils/number-utils';
 import { clamp } from 'lodash';
 import { SET_BRIGHTNESS_OPTIONS_DEFAULTS, SetBrightnessOptions } from './brightness-control-models';
+import { listen } from '@tauri-apps/api/event';
 
 @Injectable({
   providedIn: 'root',
 })
 export class SimpleBrightnessControlService {
-  private _brightness: BehaviorSubject<number> = new BehaviorSubject<number>(100);
-  private _activeTransition = new BehaviorSubject<CancellableTask | undefined>(undefined);
-  public readonly activeTransition = this._activeTransition.asObservable();
   private _advancedMode = new BehaviorSubject(false);
+  private _brightness: BehaviorSubject<number> = new BehaviorSubject<number>(100);
+  private _activeTransition = new BehaviorSubject<BrightnessTransitionTask | undefined>(undefined);
+  public readonly activeTransition = this._activeTransition.asObservable();
   private displayBrightnessDriverAvailable = false;
   public readonly advancedMode = this._advancedMode.asObservable();
 
@@ -34,6 +35,9 @@ export class SimpleBrightnessControlService {
   ) {}
 
   async init() {
+    await listen<number>('setSimpleBrightness', async (event) => {
+      await this.setBrightness(event.payload, { cancelActiveTransition: true });
+    });
     // Set brightness when switching to simple mode
     this.automationConfigService.configs
       .pipe(
@@ -74,10 +78,12 @@ export class SimpleBrightnessControlService {
   ): CancellableTask {
     const opt = { ...SET_BRIGHTNESS_OPTIONS_DEFAULTS, ...(options ?? {}) };
     if (this._brightness.value === percentage) {
-      return new CancellableTask();
+      const task = new CancellableTask();
+      task.start();
+      return task;
     }
     this._activeTransition.value?.cancel();
-    const transition = createBrightnessTransitionTask(
+    const transition = new BrightnessTransitionTask(
       'SIMPLE',
       this.setBrightness.bind(this),
       async () => this.brightness,