+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Simple mode for brightness control that consolidates image- and display brightness.
 - Brightness control dialog to directly control brightness from the main window.
+- Brightness control sliders to directly control brightness from the overlay.
 - Sleep preparation button in the overlay, to trigger automations that support this feature.
 - Sleep preparation automation for brightness control, to allow for dimming brightness levels already before going to sleep.
 - Copy buttons to the brightness automation configuration for copying current brightness levels.